few tweaks to xcode project, xcconfig to update for max 8, and gitignore.
main tweak here is to update the build location to the internal Packages folder, with a corresponding update to gitignore to ignore the mxo package. this allows us to simply create an alias to the ISF package and place in our user packages folder for easy testing of development builds.