move texture parameter handler to draw call
prior to this change, texture parameters were processed in the "param" message handler. in most cases this is fine, but makes the assumption that the param message is called with a proper GL context enabled and active, which is not an assumption we can make with jitter GL. instead we more to the draw call which is the only place that we can safely assume a valid GL context. now in the param message we simply pack the parameters into a  linklist and process that linklist in the draw call.