bind jitter textures in draw call
we must bind the texture to ensure it's properly initialized and uploaded before passing to ISF. the help file was doing this previously by sending the textures to a videoplane, however this is the proper way.

also, we add some more flags to get rid of unused/unneeded attributes, and slight tweak to audio example.