perf improvement, updated submodule, version bump
updated VVISF-GL submodule to latest version :: this is probably the 1.0.2 version of jit.gl.isf :: significant perf improvement on macs running the GL3 engine.  prior to this commit, jit.gl.isf would always try to compile passed ISFs against GL2- even if the host environment was running GL3+.  this mismatch would force macs to use IOSurfaces, which caused a significant drop in performance.  this commit makes jit.gl.isf try to compile passed ISFs against the host-version of GL, and will only try to use "the other" version of GL if the ISF cannot be compiled in the host's environment (so IOSurfaces will only be used as a last resort).